+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.2.2](https://github.com/sagawrr/react-native-nitro-vision-kit/compare/v0.2.1...v0.2.2) (2026-07-12)
+
+Photos and cutouts stay upright.
+
+* Bake EXIF orientation when loading images (iOS + Android)
+* Stop flipping cutout pixels after Vision renders on iOS
+* Fix invalid `dispose()` override on iOS segmentation results
+
 ## [0.2.1](https://github.com/sagawrr/react-native-nitro-vision-kit/compare/v0.1.1...v0.2.1) (2026-07-04)
 
 ### Features